\name{adwords}
\alias{adwords}
\docType{data}
\alias{adwords}
\title{Clicks and Impressions of a Google AdWords campaign}
\description{
  This data set contains daily impresssions and clicks of an anonymous
  Google  AdWords campaign related to an outdoor shop over the period of
  roughly two years. This data has been provided by Crealytics GmbH,
  Germany.
}
\usage{data("adwords")}
\format{A data frame containing the variables \code{id} (an identifier
  for the keywords), \code{date}, \code{keyword_string} (the string for
  advertised keywords), \code{impressions} (the number of displayed ads)
  and \code{clicks} (the daily amount of clicks).
}
\source{Crealytics GmbH.}
\references{
  Liebl, D., Rameseder, S., Rust, C. (2017) Functional Insights into
  Google Adwords. Available from the authors on request.
}
\keyword{datasets}
