% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calSecDerNatSpline.R
\name{calSecDerNatSpline}
\alias{calSecDerNatSpline}
\title{calSecDerNatSpline}
\usage{
calSecDerNatSpline(grd)
}
\arguments{
\item{grd}{A vector of length \eqn{p} indicating the grd values where the
spline functions are evaluated.}
}
\value{
A list with entries
\item{A_m}{
  The \eqn{p \times p} matrix \eqn{A} as described in (Crambes, Kneip and Sarda,
  2016)
}
\item{X_B}{
  A \eqn{p \times p} matrix of natural cubic spline basis evaluated at \code{grd}.
}
}
\description{
Natural Spline Basis and the Regularization Matrix \code{A_m}
}
\details{
\code{calSecDerNatSpline} calculates evaluated natural cubic splines
and the matrix \code{A_m} as described in Crambes, Kneip and Sarda
(2016).


\code{calSecDerNatSpline} calculates an evaluated natural cubic
spline basis corresponding to \code{grd} and a \eqn{p \times p} matrix
used for regularization in the smoothing splines estimator for the
functional linear model (Crambes, Kneip and Sarda,2016). \code{A_m} is
composed of a non-classical projection matrix and the classical
regularization matrix of squared second derivatives.
}
\examples{
\dontrun{
library(FunRegPoI)

## Define Parameters for Simulation
DGP_name    <- "Easy"
k_seq       <- c(1,seq(2, 60, 4))
error_sd    <- 0.125
N           <- 500
p           <- 300
domain      <- c(0,1)
grd         <- seq(domain[1],domain[2],length.out = p)

set.seed(1)
## simulate some data:
PoISim      <- FunRegPoISim(N = N , grd, DGP = DGP_name , error_sd = error_sd)

NaturalSplines <- calSecDerNatSpline(grd)

# PESES
EstPeses  <- FunRegPoI(Y = PoISim$Y , X_mat = PoISim$X , grd, estimator = "PESES",
                       k_seq = k_seq , A_m = NaturalSplines$A_m , X_B = NaturalSplines$X_B)

}

}
\references{
Crambes, C., Kneip, A., Sarda, P. (2009) Smoothing Splines
Estimators for Functional Linear Regression. The Annals of
Statistics, \emph{37}(1), 35-72.
}
\seealso{
FunRegPoI
}
\author{
Dominik Liebl, Stefan Rameseder and Christoph Rust
}
\keyword{models}
